+++ src/main/java/com/app/controllers/MemberController.java
@@ -1,6 +1,8 @@
 package com.app.controllers;
 
 import com.app.models.*;
+import com.app.models.exceptions.CouldNotLoadMemberException;
+import com.app.models.exceptions.MemberNotFoundException;
 import com.app.models.services.ConfigService;
 import com.app.models.services.MemberService;
 import com.app.models.services.PaymentRequestService;
