+++ src/main/java/com/app/models/DistanceModel.java
@@ -20,3 +20,4 @@ public enum DistanceModel {
     return METERS;
   }
 }
+
