+++ src/main/java/com/app/models/exceptions/MemberNotFoundException.java
@@ -1,4 +1,4 @@
-package com.app.models;
+package com.app.models.exceptions;
 
 /** Exception to be thrown when a Member is not found during lookup */
 public class MemberNotFoundException extends Exception {
